Rename TearOut file to ResizeWrapper
